Extract shared streak logic and optimize graph styling
