+++ src/ContributionGraph.svelte
@@ -1,9 +1,14 @@
 <script>
-	import {debugLog, isValidCSSColor} from './utils'
+	import {
+		debugLog,
+		isValidCSSColor,
+		buildGraphCellStyle,
+		computeHabitRenderedDays,
+	} from './utils'
 
 	import {onDestroy} from 'svelte'
 	import {parseYaml, TFile} from 'obsidian'
-	import {parseISO, format, startOfWeek, addDays, isBefore, isAfter, isSameDay, differenceInCalendarDays, isToday} from 'date-fns'
+	import {parseISO, format, startOfWeek, addDays, isBefore, isAfter, isSameDay, isToday} from 'date-fns'
 
 	export let app
 	export let name
@@ -19,6 +24,8 @@
 	let frontmatter = {}
 	let habitName = name
 	let customColor = ''
+	let fillToPreviousMonday = true
+	let cellStyle = ''
 	let savingChanges = false
 	// Keep graph streak badges aligned with default mode behavior (only meaningful multi-day streaks).
 	const STREAK_BADGE_THRESHOLD = 2
@@ -33,14 +40,22 @@
 		}
 	}
 
+	$: fillToPreviousMonday =
+		userSettings.fillToPreviousMonday !== undefined
+			? userSettings.fillToPreviousMonday
+			: globalSettings.fillToPreviousMonday
+
+	$: cellStyle = customColor ? buildGraphCellStyle(customColor) : ''
+
 	// Build the contribution graph grid: rows = days of week (Mon-Sun), columns = weeks
 	$: graph = (() => {
 		if (!dates || dates.length === 0) return {weeks: [], monthLabels: []}
 		const firstDate = parseISO(dates[0])
 		const lastDate = parseISO(dates[dates.length - 1])
 
-		// Start from the Monday of the week containing the first date
-		const weekStart = startOfWeek(firstDate, {weekStartsOn: 1}) // 1 = Monday
+		const weekStart = fillToPreviousMonday
+			? startOfWeek(firstDate, {weekStartsOn: 1})
+			: firstDate
 
 		const weeks = []
 		const monthStarts = []
@@ -107,85 +122,7 @@
 
 	$: renderedDates = (() => {
 		const maxGap = Number(frontmatter.maxGap) || 0
-		const entrySet = new Set(entries)
-
-		const days = dates.map((date) => ({
-			date,
-			ticked: entrySet.has(date),
-			gap: false,
-			deadline: false,
-			streakEnd: false,
-			streakCount: 0,
-		}))
-
-		for (const day of days) {
-			if (day.ticked || maxGap === 0) continue
-			const parsed = parseISO(day.date)
-			for (let i = 0; i < entries.length - 1; i++) {
-				const prev = parseISO(entries[i])
-				const next = parseISO(entries[i + 1])
-				if (
-					differenceInCalendarDays(parsed, prev) > 0 &&
-					differenceInCalendarDays(next, parsed) > 0
-				) {
-					if (differenceInCalendarDays(next, prev) - 1 <= maxGap) {
-						day.gap = true
-					}
-					break
-				}
-			}
-		}
-
-		const closeStreakAt = (endIdx) => {
-			let lastTickDate = null
-			for (let j = streakStartIdx; j <= endIdx; j++) {
-				if (days[j].ticked) lastTickDate = days[j].date
-			}
-
-			let count = 0
-			if (lastTickDate) {
-				const anchorIdx = entries.indexOf(lastTickDate)
-				if (anchorIdx !== -1) {
-					count = 1
-					for (let j = anchorIdx; j > 0; j--) {
-						const gapDays =
-							differenceInCalendarDays(parseISO(entries[j]), parseISO(entries[j - 1])) - 1
-						if (gapDays > maxGap) break
-						count++
-					}
-				}
-			}
-
-			days[endIdx].streakEnd = true
-			days[endIdx].streakCount = count
-		}
-
-		let streakStartIdx = -1
-		for (let i = 0; i < days.length; i++) {
-			const inStreak = days[i].ticked || days[i].gap
-			if (inStreak && streakStartIdx === -1) {
-				streakStartIdx = i
-			} else if (!inStreak && streakStartIdx !== -1) {
-				closeStreakAt(i - 1)
-				streakStartIdx = -1
-			}
-		}
-		if (streakStartIdx !== -1) {
-			closeStreakAt(days.length - 1)
-		}
-
-		if (maxGap > 0 && entries.length > 0) {
-			const today = format(new Date(), 'yyyy-MM-dd')
-			const lastEntry = entries[entries.length - 1]
-			const deadlineDate = format(
-				addDays(parseISO(lastEntry), maxGap + 1),
-				'yyyy-MM-dd',
-			)
-			if (deadlineDate >= today) {
-				const ghostDay = days.find((d) => d.date === deadlineDate)
-				if (ghostDay && !ghostDay.ticked) ghostDay.deadline = true
-			}
-		}
+		const days = computeHabitRenderedDays({dates, entries, maxGap})
 
 		return {days, byDate: new Map(days.map((day) => [day.date, day]))}
 	})()
@@ -278,58 +215,6 @@
 		app.vault.offref(modifyRef)
 	})
 
-	const getCellStyle = (cell) => {
-		if (!customColor || !(cell.ticked || cell.gap || cell.deadline || cell.today)) {
-			return ''
-		}
-
-		const toRgb = (color) => {
-			if (typeof document === 'undefined') return null
-			const canvas = document.createElement('canvas')
-			const context = canvas.getContext('2d')
-			if (!context) return null
-			context.fillStyle = '#000000'
-			context.fillStyle = color
-			const normalizedColor = context.fillStyle
-			const hexMatch = normalizedColor.match(/^#([0-9a-f]{3,8})$/i)
-			if (hexMatch) {
-				const hex = hexMatch[1]
-				if (hex.length === 3) {
-					return {
-						r: Number.parseInt(hex[0] + hex[0], 16),
-						g: Number.parseInt(hex[1] + hex[1], 16),
-						b: Number.parseInt(hex[2] + hex[2], 16),
-					}
-				}
-				if (hex.length >= 6) {
-					return {
-						r: Number.parseInt(hex.slice(0, 2), 16),
-						g: Number.parseInt(hex.slice(2, 4), 16),
-						b: Number.parseInt(hex.slice(4, 6), 16),
-					}
-				}
-				return null
-			}
-			const rgbMatch = normalizedColor.match(
-				/^rgba?\(\s*(\d+)\s*,\s*(\d+)\s*,\s*(\d+)/i,
-			)
-			if (!rgbMatch) return null
-			return {
-				r: Number.parseInt(rgbMatch[1], 10),
-				g: Number.parseInt(rgbMatch[2], 10),
-				b: Number.parseInt(rgbMatch[3], 10),
-			}
-		}
-
-		const textColor = (() => {
-			const rgb = toRgb(customColor)
-			if (!rgb) return 'var(--text-on-accent, var(--text-normal))'
-			const luminance = (0.299 * rgb.r + 0.587 * rgb.g + 0.114 * rgb.b) / 255
-			return luminance > 0.72 ? 'rgb(34, 38, 43)' : 'rgb(244, 247, 250)'
-		})()
-
-		return `--graph-cell-color: ${customColor}; --graph-today-color: ${customColor}; --graph-cell-contrast-color: color-mix(in srgb, ${customColor} 18%, black 82%); --graph-cell-text-color: ${textColor}`
-	}
 </script>
 
 <div class="contribution-graph">
@@ -365,7 +250,7 @@
 							class:contribution-graph__cell--today={cell.today}
 							class:contribution-graph__cell--streak-end={cell.streakEnd}
 							class:contribution-graph__cell--empty={!cell.isInRange}
-							style={getCellStyle(cell)}
+							style={cell.ticked || cell.gap || cell.deadline || cell.today ? cellStyle : ''}
 							title={cell.isInRange ? cell.date : ''}
 							on:click={() => cell.isInRange && toggleHabit(cell.date)}
 						>
