+++ src/Habit.svelte
@@ -1,10 +1,8 @@
 <script>
-	import {debugLog, isValidCSSColor} from './utils'
+	import {debugLog, isValidCSSColor, getDayOfTheWeek, computeHabitRenderedDays} from './utils'
 
 	import {onDestroy} from 'svelte'
 	import {parseYaml, TFile} from 'obsidian'
-	import {getDayOfTheWeek} from './utils'
-	import {differenceInCalendarDays, parseISO, format} from 'date-fns'
 
 	export let app
 	export let name
@@ -38,145 +36,15 @@
 
 	$: renderedDates = (() => {
 		const maxGap = Number(frontmatter.maxGap) || 0
-		const entrySet = new Set(entries)
 		const gapStyle =
 			userSettings.gapStyle !== undefined
 				? userSettings.gapStyle
 				: globalSettings.gapStyle
 
-		// Pass 1 — mark each date
-		const days = dates.map((date) => {
-			const ticked = entrySet.has(date)
-			let gap = false
-			if (!ticked && maxGap > 0) {
-				// Gap only between consecutive entries whose gap ≤ maxGap
-				const parsed = parseISO(date)
-				for (let i = 0; i < entries.length - 1; i++) {
-					const prev = parseISO(entries[i])
-					const next = parseISO(entries[i + 1])
-					if (
-						differenceInCalendarDays(parsed, prev) > 0 &&
-						differenceInCalendarDays(next, parsed) > 0
-					) {
-						if (differenceInCalendarDays(next, prev) - 1 <= maxGap) {
-							gap = true
-						}
-						break
-					}
-				}
-			}
-			return {
-				date,
-				ticked,
-				gap,
-				deadline: false,
-				title: '',
-				streakStart: false,
-				streakEnd: false,
-				streakCount: 0,
-				classes: '',
-			}
-		})
-
-		// Pass 2 — identify streak boundaries and counts
-		let streakStartIdx = -1
-		for (let i = 0; i <= days.length; i++) {
-			const inStreak = i < days.length && (days[i].ticked || days[i].gap)
-			if (inStreak && streakStartIdx === -1) {
-				streakStartIdx = i
-			} else if (!inStreak && streakStartIdx !== -1) {
-				// Streak just ended at i-1
-				const endIdx = i - 1
-
-				// Find first and last ticked dates in this visible run
-				let firstTickDate = null
-				let lastTickDate = null
-				for (let j = streakStartIdx; j <= endIdx; j++) {
-					if (days[j].ticked) {
-						if (!firstTickDate) firstTickDate = days[j].date
-						lastTickDate = days[j].date
-					}
-				}
-
-				// streakStart: only if the streak truly begins here
-				// (no entry within maxGap before the first visible date)
-				if (firstTickDate) {
-					const firstTickIdx = entries.indexOf(firstTickDate)
-					const prevEntry = firstTickIdx > 0 ? entries[firstTickIdx - 1] : null
-					const continuesFromBefore =
-						prevEntry &&
-						differenceInCalendarDays(
-							parseISO(firstTickDate),
-							parseISO(prevEntry),
-						) -
-							1 <=
-							maxGap
-					if (!continuesFromBefore) {
-						days[streakStartIdx].streakStart = true
-					}
-				} else {
-					days[streakStartIdx].streakStart = true
-				}
-
-				// streakEnd: only if the streak truly ends within the visible range
-				if (lastTickDate) {
-					const lastTickIdx = entries.indexOf(lastTickDate)
-					const nextEntry =
-						lastTickIdx < entries.length - 1 ? entries[lastTickIdx + 1] : null
-					const continuesAfter =
-						nextEntry &&
-						differenceInCalendarDays(
-							parseISO(nextEntry),
-							parseISO(lastTickDate),
-						) -
-							1 <=
-							maxGap
-					if (!continuesAfter) {
-						days[endIdx].streakEnd = true
-					}
-				} else {
-					days[endIdx].streakEnd = true
-				}
-
-				// Count: walk backward through entries from the last visible tick
-				let count = 0
-				if (lastTickDate) {
-					const anchorIdx = entries.indexOf(lastTickDate)
-					if (anchorIdx !== -1) {
-						count = 1
-						for (let j = anchorIdx; j > 0; j--) {
-							const gapDays =
-								differenceInCalendarDays(
-									parseISO(entries[j]),
-									parseISO(entries[j - 1]),
-								) - 1
-							if (gapDays > maxGap) break
-							count++
-						}
-					}
-				}
-
-				days[endIdx].streakCount = count
-
-				streakStartIdx = -1
-			}
-		}
-
-		// Pass 3 — ghost dot on the last day of the gap (deadline to keep streak alive)
-		if (maxGap > 0 && entries.length > 0) {
-			const today = format(new Date(), 'yyyy-MM-dd')
-			const lastEntry = entries[entries.length - 1]
-			const deadlineDate = format(
-				new Date(parseISO(lastEntry).getTime() + (maxGap + 1) * 86400000),
-				'yyyy-MM-dd',
-			)
-			if (deadlineDate >= today) {
-				const ghostDay = days.find((d) => d.date === deadlineDate)
-				if (ghostDay && !ghostDay.ticked) {
-					ghostDay.deadline = true
-				}
-			}
-		}
+		const days = computeHabitRenderedDays({dates, entries, maxGap}).map((day) => ({
+			...day,
+			classes: '',
+		}))
 
 		// Build classes
 		for (const day of days) {
